refactor(tui): colocate Screen::List into tui/screens/list.rs
